requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-nmap==0.7.1